Add search indexing

+++ backend/seed.js
@@ -6,6 +6,7 @@ const { fileURLToPath } = require('url');
 const { dirname, join, extname } = require('path');
 
 const dbName = 'courseEvaluations';
+const collectionName = 'evaluations';
 const dataFilesPath = join(__dirname, '/data');
 const metadataTokens = [
   { property: 'term', parse: (rawObj) => rawObj.metadata['Term'] },
@@ -23,6 +24,23 @@ const metadataTokens = [
   },
 ];
 
+// creates N-grams of each search token starting from 2
+const createSearchIndexes = (...args) => {
+  const str = args.map((arg) => arg.trim().replace(/\s\s+/g, ' ')).join(' ');
+  const min = 2;
+
+  return Array.from(
+    str.split(' ').reduce((set, token) => {
+      if (token.length >= min) {
+        for (let i = min; i <= str.length && i <= token.length; ++i) {
+          set.add(token.substring(0, i));
+        }
+      }
+      return set;
+    }, new Set()),
+  );
+};
+
 const seedDB = async () => {
   const client = new MongoClient(process.env.ATLAS_URI);
 
@@ -30,10 +48,14 @@ const seedDB = async () => {
     await client.connect();
     console.log('Connected to the server...');
 
-    const collection = client.db(dbName).collection('evaluations');
+    const db = client.db(dbName);
+    const collection = client.db(dbName).collection(collectionName);
+    const collectionInfo = await db.listCollections({ name: collectionName }).next();
 
     // remove the entire collection before seeding it once again
-    await collection.drop();
+    if (collectionInfo) {
+      await collection.drop();
+    }
 
     const entries = [];
     const fileNames = await fs.readdir(dataFilesPath);
@@ -50,6 +72,10 @@ const seedDB = async () => {
           course.metadata = rawCourse.metadata;
           metadataTokens.forEach(({ property, parse }) => (course[property] = parse(rawCourse)));
           course.questionSections = rawCourse.questionSections;
+          course.searchIndexes = createSearchIndexes(
+            course.description,
+            course.instructors.join(' '),
+          );
           entries.push(course);
         });
       }
